Ability to hide any click (on the DailyActivity) On the Overview activity I switched the name of the day with the calendar day
All queries and calculations have been updated to only take into account the ‘unmarked’ data
On the Overview -> [!] used
On the Last 7 Lays -> ! Used (the brackets have been removed so that the text with the ! mark fit)
On the Month view -> the discussed dot
Here I’ve also added a gray day with a dot for when we only have hidden data
On the Hourly view (see below what this is) -> ! used
Navigation menus have been updated to allow easier navigation between views.
An Hourly Activity has been added which depicts the ups & downs per hour (for all days)
An analytics table has been added to record the action of hiding/unhiding a click that records the result of the hide/unhide operation.